Feat/partial run testflow implementation

+++ packages/@sparrow-workspaces/src/features/testflow-explorer/layout/TestflowExplorer.svelte
@@ -253,7 +253,7 @@
             "onCheckEdges",
             "onClick",
             "onOpenAddCustomRequestModal",
-            "onOpenDeleteModal",
+            "onContextMenu",
             "onUpdateSelectAPI",
             "onOpenSaveNodeRequestModal",
           ];
@@ -526,7 +526,7 @@
     // handles run from from start button click
     if (_id === "0") {
       await onClickRun();
-      selectFirstNode();
+      selectNode("2");
       MixpanelEvent(Events.Run_TestFlows);
       return;
     }
@@ -581,8 +581,15 @@
                 folderId,
               );
             },
-            onOpenDeleteModal: function (id: string) {
-              handleDeleteModal(id);
+            onContextMenu: function (id: string, _event: string) {
+              if (_event === "delete") {
+                handleDeleteModal(id);
+              } else if (
+                _event === "run-from-here" ||
+                _event === "run-till-here"
+              ) {
+                partialRun(id, _event);
+              }
             },
             onOpenAddCustomRequestModal: function (id: string) {
               handleOpenAddCustomRequestModal();
@@ -674,8 +681,15 @@
                 folderId,
               );
             },
-            onOpenDeleteModal: function (id: string) {
-              handleDeleteModal(id);
+            onContextMenu: function (id: string, _event: string) {
+              if (_event === "delete") {
+                handleDeleteModal(id);
+              } else if (
+                _event === "run-from-here" ||
+                _event === "run-till-here"
+              ) {
+                partialRun(id, _event);
+              }
             },
             onOpenAddCustomRequestModal: function (id: string) {
               handleOpenAddCustomRequestModal();
@@ -911,10 +925,10 @@
   /**
    * Select all the existing nodes
    */
-  const selectFirstNode = () => {
+  const selectNode = (_id: string) => {
     nodes.update((_nodes: Node[] | any[]) => {
-      _nodes.forEach((_nodeItem, index) => {
-        if (index === 1) {
+      _nodes.forEach((_nodeItem) => {
+        if (_nodeItem.id === _id) {
           _nodeItem.selected = true;
         } else {
           _nodeItem.selected = false;
@@ -978,6 +992,14 @@
       sampleApiData = onRunSampleApi();
     }, 0);
   });
+
+  const partialRun = async (_id: string, _event: string) => {
+    if (!testflowStore?.isTestFlowRunning) {
+      unselectNodes();
+      await onClickRun(_id, _event);
+      selectNode(_id);
+    }
+  };
 </script>
 
 <div
@@ -1033,7 +1055,7 @@
               onClick={async () => {
                 unselectNodes();
                 await onClickRun();
-                selectFirstNode();
+                selectNode("2");
                 MixpanelEvent(Events.Run_TestFlows);
               }}
             />
